requirements.txt
dnspython>=2,<3